+++ appliance_settings.go
@@ -41,6 +41,7 @@ type ApplianceSettings struct {
 		ID   int64  `json:"id"`
 		Name string `json:"name"`
 	} `json:"enabledZoneTypes"`
+	StatsRetainmentPeriod int64 `json:"statsRetainmentPeriod"`
 }
 
 type GetApplianceSettingsResult struct {