+++ instances.go
@@ -12,28 +12,118 @@ var (
 
 // Instance structures for use in request and response payloads
 type Instance struct {
-	ID           int64                  `json:"id"`
-	Name         string                 `json:"name"`
-	Description  string                 `json:"description"`
-	InstanceType map[string]interface{} `json:"instanceType"`
-	Layout       map[string]interface{} `json:"layout"`
-	Group        map[string]interface{} `json:"group"`
-	Cloud        map[string]interface{} `json:"cloud"`
-	Environment  string                 `json:"instanceContext"`
-	Plan         InstancePlan           `json:"plan"`
-	Config       map[string]interface{} `json:"config"`
-	Labels       []string               `json:"labels"`
-	Version      string                 `json:"instanceVersion"`
-	Status       string                 `json:"status"`
-	Owner        Owner                  `json:"owner"`
-
-	// might want to define types for these too
-	Volumes              *[]map[string]interface{} `json:"volumes"`
-	Interfaces           *[]map[string]interface{} `json:"interfaces"`
-	Controllers          *[]map[string]interface{} `json:"controllers"`
-	Tags                 *[]map[string]interface{} `json:"tags"`
-	Metadata             *[]map[string]interface{} `json:"metadata"`
-	EnvironmentVariables *[]map[string]interface{} `json:"evars"`
+	ID        int64  `json:"id"`
+	UUID      string `json:"uuid"`
+	AccountId int64  `json:"accountId"`
+	Tenant    struct {
+		ID   int64  `json:"id"`
+		Name string `json:"name"`
+	} `json:"tenant"`
+	Name         string `json:"name"`
+	Description  string `json:"description"`
+	InstanceType struct {
+		ID       int64  `json:"id"`
+		Name     string `json:"name"`
+		Code     string `json:"code"`
+		Category string `json:"category"`
+		Image    string `json:"image"`
+	} `json:"instanceType"`
+	Layout struct {
+		ID                int64  `json:"id"`
+		Name              string `json:"name"`
+		ProvisionTypeId   int64  `json:"provisionTypeId"`
+		ProvisionTypeCode string `json:"provisionTypeCode"`
+	} `json:"layout"`
+	Group struct {
+		ID   int64  `json:"id"`
+		Name string `json:"name"`
+	} `json:"group"`
+	Cloud struct {
+		ID   int64  `json:"id"`
+		Name string `json:"name"`
+	} `json:"cloud"`
+	Containers []int64 `json:"containers"`
+	Servers    []int64 `json:"servers"`
+	Resources  []struct {
+		ID              int64    `json:"id"`
+		UUID            string   `json:"uuid"`
+		Code            string   `json:"code"`
+		Category        string   `json:"category"`
+		Name            string   `json:"name"`
+		DisplayName     string   `json:"displayName"`
+		Labels          []string `json:"labels"`
+		ResourceVersion string   `json:"resourceVersion"`
+		ResourceContext string   `json:"resourceContext"`
+		Owner           struct {
+			Id   int64  `json:"id"`
+			Name string `json:"name"`
+		} `json:"owner"`
+		ResourceType string `json:"resourceType"`
+		ResourceIcon string `json:"resourceIcon"`
+		Type         struct {
+			Id   int64  `json:"id"`
+			Code string `json:"code"`
+			Name string `json:"name"`
+		} `json:"type"`
+		Status     string `json:"status"`
+		Enabled    bool   `json:"enabled"`
+		ExternalId string `json:"externalId"`
+	} `json:"resources"`
+	ConnectionInfo []struct {
+		Ip   string `json:"ip"`
+		Port int64  `json:"port"`
+	} `json:"connectionInfo"`
+	Environment string                   `json:"environment"`
+	Plan        InstancePlan             `json:"plan"`
+	Config      map[string]interface{}   `json:"config"`
+	Labels      []string                 `json:"labels"`
+	Version     string                   `json:"instanceVersion"`
+	Status      string                   `json:"status"`
+	Owner       Owner                    `json:"owner"`
+	Volumes     []map[string]interface{} `json:"volumes"`
+	Interfaces  []map[string]interface{} `json:"interfaces"`
+	Controllers []map[string]interface{} `json:"controllers"`
+	Tags        []struct {
+		Id    int64  `json:"id"`
+		Name  string `json:"name"`
+		Value string `json:"value"`
+	} `json:"tags"`
+	Metadata             []map[string]interface{} `json:"metadata"`
+	EnvironmentVariables []struct {
+		Name   string `json:"name"`
+		Value  string `json:"value"`
+		Export bool   `json:"export"`
+		Masked bool   `json:"masked"`
+	} `json:"evars"`
+	CustomOptions  map[string]interface{} `json:"customOptions"`
+	MaxMemory      int64                  `json:"maxMemory"`
+	MaxStorage     int64                  `json:"maxStorage"`
+	MaxCores       int64                  `json:"maxCores"`
+	CoresPerSocket int64                  `json:"coresPerSocket"`
+	MaxCpu         int64                  `json:"maxCpu"`
+	HourlyCost     float64                `json:"hourlyCost"`
+	HourlyPrice    float64                `json:"hourlyPrice"`
+	InstancePrice  struct {
+		Price    float64 `json:"price"`
+		Cost     float64 `json:"cost"`
+		Currency string  `json:"currency"`
+		Unit     string  `json:"unit"`
+	} `json:"instancePrice"`
+	DateCreated       string `json:"dateCreated"`
+	LastUpdated       string `json:"lastUpdated"`
+	HostName          string `json:"hostName"`
+	DomainName        string `json:"domainName"`
+	EnvironmentPrefix string `json:"environmentPrefix"`
+	FirewallEnabled   bool   `json:"firewallEnabled"`
+	NetworkLevel      string `json:"networkLevel"`
+	AutoScale         bool   `json:"autoScale"`
+	InstanceContext   string `json:"instanceContext"`
+	Locked            bool   `json:"locked"`
+	IsScalable        bool   `json:"isScalable"`
+	CreatedBy         struct {
+		Id       int64  `json:"id"`
+		Username string `json:"username"`
+	} `json:"createdBy"`
 }
 
 type InstancePlan struct {
@@ -45,6 +135,52 @@ type InstancePlan struct {
 type Owner struct {
 	ID       int64  `json:"id"`
 	Username string `json:"username"`
+}
+
+type ContainerDetails struct {
+	ID               int64  `json:"id"`
+	UUID             string `json:"uuid"`
+	Name             string `json:"name"`
+	IP               string `json:"ip"`
+	InternalIp       string `json:"internalIp"`
+	InternalHostname string `json:"internalHostname"`
+	ExternalHostname string `json:"externalHostname"`
+	ExternalDomain   string `json:"externalDomain"`
+	ExternalFqdn     string `json:"externalFqdn"`
+	AccountId        int64  `json:"accountId"`
+	Instance         struct {
+		ID   int64  `json:"id"`
+		Name string `json:"name"`
+	} `json:"instance"`
+	ContainerType struct {
+		ID       int64  `json:"id"`
+		Name     string `json:"name"`
+		Code     string `json:"code"`
+		Category string `json:"category"`
+	} `json:"containerType"`
+	Server struct {
+		ID               int64  `json:"id"`
+		UUID             string `json:"uuid"`
+		ExternalId       string `json:"externalId"`
+		InternalId       string `json:"internalId"`
+		ExternalUniqueId string `json:"externalUniqueId"`
+		Name             string `json:"name"`
+		ExternalName     string `json:"externalName"`
+		Hostname         string `json:"hostname"`
+		AccountId        int64  `json:"accountId"`
+		SshHost          string `json:"sshHost"`
+		ExternalIp       string `json:"externalIp"`
+		InternalIp       string `json:"internalIp"`
+		Platform         string `json:"platform"`
+		PlatformVersion  string `json:"platformVersion"`
+		AgentInstalled   bool   `json:"agentInstalled"`
+		AgentVersion     string `json:"agentVersion"`
+		SourceImage      struct {
+			ID   int64  `json:"id"`
+			Name string `json:"name"`
+			Code string `json:"code"`
+		} `json:"sourceImage"`
+	} `json:"server"`
 }
 
 // ListInstancesResult structure parses the list instances response payload
@@ -165,7 +301,7 @@ func (client *Client) ListInstancePlans(req *Request) (*Response, error) {
 	})
 }
 
-//todo: need this api endpoint still, and consolidate to /api/plans perhaps
+// todo: need this api endpoint still, and consolidate to /api/plans perhaps
 func (client *Client) GetInstancePlan(id int64, req *Request) (*Response, error) {
 	return client.Execute(&Request{
 		Method:      "GET",